Remove UUID from GameState and relocate deserializer
Remove the UUID field from GameState (no longer generated/stored) and update tests accordingly to stop asserting on the UUID. Move/rename UUIDDerializer into nomadrealms.math.id.serialization and clean up its imports. Test GameStateDerializerTest adjusted to remove UUID setup/assertion; serialization/deserializer behavior for other fields remains covered.
